route_importer_web/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
flask==2.3.3
flask-cors==4.0.0
google-generativeai
python-dotenv
dashscope
